Migrate doc-gen build to Rolldown
